requirements/base.txt
starlette
uvicorn
tortoise-orm
aiosqlite
graphene
aiofiles
pychord
